requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
anthropic
python-dotenv
httpx
redis
slowapi
textblob
regex
phonenumbers
validators
google-generativeai
pytest
pytest-asyncio
requests
